added business logic to project

+++ src/main/java/com/geoffrey/mini_trello_back/project/ProjectService.java
@@ -1,5 +1,7 @@
 package com.geoffrey.mini_trello_back.project;
 
+import com.geoffrey.mini_trello_back.auth.AuthUtils;
+import com.geoffrey.mini_trello_back.auth.exceptions.AccessDeniedException;
 import com.geoffrey.mini_trello_back.common.ResponsePaginatedDto;
 import com.geoffrey.mini_trello_back.common.ResponsePaginatedMapper;
 import com.geoffrey.mini_trello_back.profile.Profile;
@@ -12,14 +14,17 @@ import com.geoffrey.mini_trello_back.project.dto.PatchProjectDto;
 import com.geoffrey.mini_trello_back.project.dto.PatchProjectOwnerDto;
 import com.geoffrey.mini_trello_back.project.dto.ProjectResponseDto;
 import com.geoffrey.mini_trello_back.project.exceptions.ProjectNotFoundException;
+import com.geoffrey.mini_trello_back.task.Task;
 import com.geoffrey.mini_trello_back.task.TaskMapper;
 import com.geoffrey.mini_trello_back.task.TaskRepository;
 import com.geoffrey.mini_trello_back.task.dto.SimpleTaskResponseDto;
+import com.geoffrey.mini_trello_back.user.User;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 
 @Service
 public class ProjectService {
@@ -46,27 +51,38 @@ public class ProjectService {
         this.responsePaginatedMapper = responsePaginatedMapper;
     }
 
-    public ResponsePaginatedDto<List<ProjectResponseDto>> listProjects(Pageable pageable) {
-        Page<ProjectResponseDto> page = projectRepository.findAll(pageable).map(projectMapper::toProjectResponseDto);
+    public ResponsePaginatedDto<List<ProjectResponseDto>> listProjects(Pageable pageable, User currentUser) {
+        Profile profile = AuthUtils.getProfileFromUser(currentUser);
+        List<Integer> projectIds = projectRepository.findProjectIdsByProfileId(profile.getId());
+        Page<ProjectResponseDto> page = projectRepository.findProjectsRelatedByProfileId(projectIds, profile.getId(), pageable)
+                .map(projectMapper::toProjectResponseDto);
         return responsePaginatedMapper.toResponsePaginatedDto(page, pageable);
     }
 
-    public ProjectResponseDto createProject(CreateProjectDto projectDto) {
-        Integer profileId = projectDto.profileId();
-        Profile owner = profileRepository.findById(profileId).orElseThrow(() -> new ProfileNotFoundException((profileId)));
-        Project project = projectMapper.toProject(projectDto, owner);
+    public ProjectResponseDto createProject(CreateProjectDto projectDto, User currentUser) {
+        Profile profile = AuthUtils.getProfileFromUser(currentUser);
+        Project project = projectMapper.toProject(projectDto, profile);
         Project newProject = projectRepository.save(project);
         return projectMapper.toProjectResponseDto(newProject);
     }
 
-    public ProjectResponseDto getProject(Integer projectId) {
-        return projectRepository.findById(projectId)
-                .map(projectMapper::toProjectResponseDto)
-                .orElseThrow(() -> new ProjectNotFoundException(projectId));
+    public ProjectResponseDto getProject(Integer projectId, User currentUser) {
+        Profile profile = AuthUtils.getProfileFromUser(currentUser);
+        Project project = projectRepository.findById(projectId).orElseThrow(() -> new ProjectNotFoundException(projectId));
+
+        checkProjectRelatedToProfile(profile, project);
+
+        return projectMapper.toProjectResponseDto(project);
     }
 
-    public ProjectResponseDto patchProject(Integer projectId, PatchProjectDto projectDto) {
+
+    public ProjectResponseDto patchProject(Integer projectId, PatchProjectDto projectDto, User currentUser) {
+        Profile profile = AuthUtils.getProfileFromUser(currentUser);
         Project project = projectRepository.findById(projectId).orElseThrow(() -> new ProjectNotFoundException(projectId));
+
+        if (!Objects.equals(project.getOwner(), profile)) {
+            throw new AccessDeniedException();
+        }
 
         project.setName(projectDto.name());
         if (projectDto.description() != null) {
@@ -77,27 +93,40 @@ public class ProjectService {
         return projectMapper.toProjectResponseDto(newProject);
     }
 
-    public ProjectResponseDto patchProjectOwner(Integer projectId, PatchProjectOwnerDto projectDto) {
-        Integer profileId = projectDto.profileId();
-
-        Profile profile = profileRepository.findById(profileId).orElseThrow(() -> new ProfileNotFoundException(profileId));
+    public ProjectResponseDto patchProjectOwner(Integer projectId, PatchProjectOwnerDto projectDto, User currentUser) {
+        Profile profile = AuthUtils.getProfileFromUser(currentUser);
         Project project = projectRepository.findById(projectId).orElseThrow(() -> new ProjectNotFoundException(projectId));
+        if (!Objects.equals(project.getOwner(), profile)) {
+            throw new AccessDeniedException();
+        }
 
-        project.setOwner(profile);
+        Integer profileId = projectDto.profileId();
+        Profile newOwner = profileRepository.findById(profileId).orElseThrow(() -> new ProfileNotFoundException(profileId));
+        project.setOwner(newOwner);
 
         Project newProject = projectRepository.save(project);
         return projectMapper.toProjectResponseDto(newProject);
     }
 
-    public void deleteProject(Integer projectId) {
+    public void deleteProject(Integer projectId, User currentUser) {
+        Profile profile = AuthUtils.getProfileFromUser(currentUser);
+
         Project project = projectRepository
                 .findById(projectId)
                 .orElseThrow(() -> new ProjectNotFoundException(projectId));
+
+        if (!Objects.equals(project.getOwner(), profile)) {
+            throw new AccessDeniedException();
+        }
+
         projectRepository.delete(project);
     }
 
-    public ResponsePaginatedDto<List<SimpleTaskResponseDto>> listProjectTasks(Integer projectId, Pageable pageable) {
-        checkProject(projectId);
+    public ResponsePaginatedDto<List<SimpleTaskResponseDto>> listProjectTasks(Integer projectId, Pageable pageable, User currentUser) {
+        Profile profile = AuthUtils.getProfileFromUser(currentUser);
+        Project project = projectRepository.findById(projectId).orElseThrow(() -> new ProjectNotFoundException(projectId));
+        checkProjectRelatedToProfile(profile, project);
+
         Page<SimpleTaskResponseDto> page = taskRepository
                 .findTasksByProjectId(projectId, pageable)
                 .map(taskMapper::toSimpleTaskResponseDto);
@@ -105,18 +134,23 @@ public class ProjectService {
     }
 
 
-    public ResponsePaginatedDto<List<SimpleProfileResponseDto>> listProjectMembers(Integer projectId, Pageable pageable) {
-        checkProject(projectId);
+    public ResponsePaginatedDto<List<SimpleProfileResponseDto>> listProjectMembers(Integer projectId, Pageable pageable, User currentUser) {
+        Profile profile = AuthUtils.getProfileFromUser(currentUser);
+        Project project = projectRepository.findById(projectId).orElseThrow(() -> new ProjectNotFoundException(projectId));
+        checkProjectRelatedToProfile(profile, project);
+
         Page<SimpleProfileResponseDto> page = profileRepository
                 .findProfilesByRelatedProject(projectId, pageable)
                 .map(profileMapper::toSimpleProfileResponseDto);
         return responsePaginatedMapper.toResponsePaginatedDto(page, pageable);
     }
 
-    private void checkProject(Integer projectId) {
-        boolean exists = projectRepository.existsById(projectId);
-        if (!exists) {
-            throw new ProjectNotFoundException(projectId);
+    private void checkProjectRelatedToProfile(Profile profile, Project project) {
+        List<Integer> taskIds = profile.getTasks().stream().map(Task::getId).toList();
+        List<Integer> projectIds = projectRepository.findProjectIdsByTaskIds(taskIds);
+
+        if (!Objects.equals(project.getOwner(), profile) || !projectIds.contains(project.getId())) {
+            throw new AccessDeniedException();
         }
     }
